Change type of HSB/HSL color components from Int to Float

+++ Tests/AmazingColorHelpersTests/AmazingColorHelpersTests+Normalized+HueSaturationBrightness.swift
@@ -1,6 +1,6 @@
 //
 //  AmazingColorHelpersTests+Normalized+HueSaturationBrightness.swift
-//  
+//
 //
 //  Created by Timur Khamidov on 14.01.2022.
 //
@@ -15,9 +15,9 @@ extension AmazingColorHelpersTests {
     
     func test_amazingColorHelpers_normalized_hue_saturation_brightness_whenAllValuesAreLowerThanNormal() {
         // given
-        let hue: Int = -12
-        let saturation: Int = -23
-        let brightness: Int = -100
+        let hue: Float = -12.0
+        let saturation: Float = -23.0
+        let brightness: Float = -100.0
         
         // when
         let (nHue, nSaturation, nBrightness) = AmazingColorHelpers.shared.normalized(
@@ -27,16 +27,16 @@ extension AmazingColorHelpersTests {
         )
         
         // then
-        XCTAssertEqual(nHue, 0)
-        XCTAssertEqual(nSaturation, 0)
-        XCTAssertEqual(nBrightness, 0)
+        XCTAssertEqual(nHue, 0.0)
+        XCTAssertEqual(nSaturation, 0.0)
+        XCTAssertEqual(nBrightness, 0.0)
     }
     
     func test_amazingColorHelpers_normalized_hue_saturation_brightness_whenAllValuesAreHigherThanNormal() {
         // given
-        let hue: Int = 456
-        let saturation: Int = 123
-        let brightness: Int = 1029
+        let hue: Float = 456.0
+        let saturation: Float = 123.0
+        let brightness: Float = 1029.0
         
         // when
         let (nHue, nSaturation, nBrightness) = AmazingColorHelpers.shared.normalized(
@@ -46,16 +46,16 @@ extension AmazingColorHelpersTests {
         )
         
         // then
-        XCTAssertEqual(nHue, 0)
-        XCTAssertEqual(nSaturation, 100)
-        XCTAssertEqual(nBrightness, 100)
+        XCTAssertEqual(nHue, 0.0)
+        XCTAssertEqual(nSaturation, 100.0)
+        XCTAssertEqual(nBrightness, 100.0)
     }
     
     func test_amazingColorHelpers_normalized_hue_saturation_brightness_whenHueIsEqualTo360Degrees() {
         // given
-        let hue: Int = 360
-        let saturation: Int = 65
-        let brightness: Int = 95
+        let hue: Float = 360.0
+        let saturation: Float = 65.0
+        let brightness: Float = 95.0
         
         // when
         let (nHue, nSaturation, nBrightness) = AmazingColorHelpers.shared.normalized(
@@ -65,16 +65,16 @@ extension AmazingColorHelpersTests {
         )
         
         // then
-        XCTAssertEqual(nHue, 0)
+        XCTAssertEqual(nHue, 0.0)
         XCTAssertEqual(nSaturation, saturation)
         XCTAssertEqual(nBrightness, brightness)
     }
     
     func test_amazingColorHelpers_normalized_hue_saturation_brightness_whenAllValuesAreNormal() {
         // given
-        let hue: Int = 350
-        let saturation: Int = 65
-        let brightness: Int = 95
+        let hue: Float = 350.0
+        let saturation: Float = 65.0
+        let brightness: Float = 95.0
         
         // when
         let (nHue, nSaturation, nBrightness) = AmazingColorHelpers.shared.normalized(
